% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addGitFolders}
\alias{addGitFolders}
\title{Extent folder list by git repository}
\usage{
addGitFolders(folders, git_repository = NULL, git_subfolders = NULL)
}
\arguments{
\item{folders}{list of subfolders within the project directory.}

\item{git_repository}{name of the project's git repository. Will be
added to the folders and subfolders}

\item{git_subfolders}{subdirectories within git repository that should be
created.}
}
\description{
Extent folder list by git repository and create subdirectories
according to default values.
}
\examples{
\dontrun{

addGitFolders(folders = c("data", "data/tmp"), git_repository = "myproject")

}
}
\author{
Christoph Reudenbach, Thomas Nauss
}
\keyword{internal}
