% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addProjectTemplate.R
\name{addProjectTemplate}
\alias{addProjectTemplate}
\title{Create A Custom Project Template}
\usage{
addProjectTemplate(packageName, overwrite = FALSE)
}
\arguments{
\item{packageName}{Name of the package}
}
\value{
Used for side effects.
}
\description{
This function can be used by package developer to add their own project
template to a package they are developing.
}
