% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternativeEnvi.R
\name{alternativeEnvi}
\alias{alternativeEnvi}
\title{Define basic or alternative environment depending on actual system}
\usage{
alternativeEnvi(
  root_folder = tempdir(),
  alt_env_id = NULL,
  alt_env_value = NULL,
  alt_env_root_folder = NULL
)
}
\arguments{
\item{root_folder}{root directory of the project.}

\item{alt_env_id}{alternative system environment attribute used to
check for setting an alternative \code{root_folder}.}

\item{alt_env_value}{value of the attribute for which the alternative
root directory of the project should be set.}

\item{alt_env_root_folder}{alternative root directory.}
}
\value{
Root folder name.
}
\description{
Defines the basic environment settings depending on the
actual system environment.
}
\examples{
\dontrun{
alternativeEnvi(root_folder = tempdir(), alt_env_id = "COMPUTERNAME",
alt_env_value = "PCRZP", alt_env_root_folder = "D:\\\\BEN\\\\edu")
}
}
\seealso{
\code{\link[=createEnvi]{createEnvi()}}
}
\author{
Christoph Reudenbach, Thomas Nauss
}
