% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEnvi.R
\name{createEnvi}
\alias{createEnvi}
\title{Define and create a project environment}
\usage{
createEnvi(
  root_folder = tempdir(),
  folders = c("data", "data/tmp"),
  folder_names = NULL,
  git_repository = NULL,
  git_subfolders = c("src", "doc"),
  path_prefix = "path_",
  global = FALSE,
  libs = NULL,
  fcts_folder = NULL,
  source_functions = !is.null(fcts_folder),
  alt_env_id = NULL,
  alt_env_value = NULL,
  alt_env_root_folder = NULL,
  lut_mode = FALSE,
  create_folders = TRUE
)
}
\arguments{
\item{root_folder}{root directory of the project.}

\item{folders}{list of subfolders within the project directory.}

\item{folder_names}{names of the variables that point to subfolders. If not
provided, the base paths of the folders is used.}

\item{git_repository}{name of the project's git repository. Will be
added to the folders and subfolders defined in default "lut" or supplied by
user will be created.}

\item{git_subfolders}{subdirectories within git repository that should be
created.}

\item{path_prefix}{a prefix for the folder names.}

\item{global}{logical: export path strings as global variables?}

\item{libs}{vector with the  names of libraries}

\item{fcts_folder}{path of the folder holding the functions. All files in
this folder will be sourced.}

\item{source_functions}{logical: should functions be sourced?}

\item{alt_env_id}{alternative system environment attribute used to
check for setting an alternative \code{root_folder}.}

\item{alt_env_value}{value of the attribute for which the alternative
root directory of the project should be set.}

\item{alt_env_root_folder}{alternative root directory.}

\item{lut_mode}{use predefined environmental settings. In this case, only the
name of the git repository must be supplied to the function.}

\item{create_folders}{create folders if not existing already.}
}
\value{
A list containing the project settings.
}
\description{
Defines folder structures and creates them if necessary, loads
libraries, and sets other project relevant parameters.
}
\examples{
\dontrun{
createEnvi(root_folder = "~/edu", folders = c("data/", "data/tmp/"),
libs = c("link2GI"),
alt_env_id = "COMPUTERNAME", alt_env_value = "PCRZP",
alt_env_root_folder = "D:\\\\BEN\\\\edu")
}
}
\seealso{
\code{\link[=alternativeEnvi]{alternativeEnvi()}}
}
\author{
Christoph Reudenbach, Thomas Nauss
}
