% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnvi.R
\name{getEnvi}
\alias{getEnvi}
\title{Get directory structure as nested list}
\usage{
getEnvi(root_folder)
}
\arguments{
\item{root_folder}{root directory of the project}
}
\value{
A list containing the directory structure
}
\description{
Reproduces a system folder structure as a nested list
}
\examples{
\dontrun{

p <- getEnvi("~/project/")
p$here
p$data$here

}


}
\author{
Marvin Ludwig
}
