% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpMa.R
\name{lutInfo}
\alias{lutInfo}
\title{Get values of default environment from internal look-up table}
\usage{
lutInfo()
}
\value{
List containing lut content.
}
\description{
Get values of default environment look-up table (not required
for the package but to cross-check from a user).
}
\details{
None
}
\examples{
None
\dontrun{
lutInfo()
}
}
