% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpMa.R
\name{lutUpdate}
\alias{lutUpdate}
\title{Update values of default environment to internal look-up table}
\usage{
lutUpdate()
}
\value{
List containing lut content.
}
\description{
Update values of default environment to internal look-up table.
Run it after updating \code{\link[=pckgDefaults]{pckgDefaults()}}.
}
\details{
None
}
\examples{
None
\dontrun{
lutInfo()
}
}
