% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_build.R
\name{manuallyBuild}
\alias{manuallyBuild}
\title{CLI based build of packages}
\usage{
manuallyBuild(dsn = getwd(), pkgDir = "H:/Dokumente", document = TRUE, ...)
}
\arguments{
\item{dsn}{root directory of the package that is build by default the current wd}

\item{pkgDir}{the package library home directory}
}
\description{
CLI based build of packages in network based R environments
}
\examples{
\dontrun{
manuallyBuild()
}
}
\author{
Christoph Reudenbach, Florian Detsch
}
