% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckgDefaults.R
\name{pckgDefaults}
\alias{pckgDefaults}
\title{Create look-up table data for defining environment defaults}
\usage{
pckgDefaults()
}
\value{
A list containing the default project settings.
}
\description{
Create look-up table data for defining environment defaults
}
\details{
After changing something update environment
}
\examples{
\dontrun{
pckgDefaults()
}
}
\author{
Christoph Reudenbach, Thomas Nauss
}
\keyword{internal}
