% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{prLibrary}
\alias{prLibrary}
\title{Load and install libraries}
\usage{
prLibrary(...)
}
\arguments{
\item{...}{name of the libraries to load. The names need to be quoted.}
}
\description{
The function tries to load all libraries passed as argument. For those that
are not installed, it tries to install them and then load them.
}
\examples{
prLibrary("data.table", "plyr")

}
\seealso{
\code{\link{prSource}}
}

