% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputDefaults.R
\name{prOutputDefaults}
\alias{prOutputDefaults}
\title{Set default values of output functions}
\usage{
prOutputDefaults(...)
}
\arguments{
\item{...}{named arguments. The name of the arguments must be "table" to set the
defaults of \code{\link{prWriteTable}}, \code{\link{prWriteCsv}} or
\code{\link{prWriteCsv2}} or the name of the output function: "png" for
\code{\link{prPng}}, "pdf" for \code{\link{prDdf}}. The value of each
argument is a named list containing the new default value of the
parameters one wants to modify (see examples). To reset the defaults values
of a function, set the value to \code{NULL}.}
}
\value{
\code{prOutputDefaults} invisibly returns the list of modified defaults
  values.
}
\description{
This function can modify the default values of the parameters of output
function like \code{\link{prWriteTable}} or \code{\link{prPdf}}. This can be
useful for instance to set in one and only one location the default size
of output images, the default font...
}
\examples{
\dontrun{
# Remove row names of table output:
prOutputDefaults(table = list(row.names = FALSE))

# Modify the default size of pdf output:
prOutputDefaults(pdf = list(width = 8, height = 6))

# Reset default values for pdf and table output
prOutputDefaults(table = NULL, pdf = NULL)
}

}

