% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{prPath}
\alias{prPath}
\title{Get the full path of a file}
\usage{
prPath(x)
}
\arguments{
\item{x}{Relative path of a file contained in the project.}
}
\value{
Absolute path of the file.
}
\description{
The function returns the absolute path of a file contained in the current
project.
}
\examples{
dontrun{
  prPath("data")
}

}

