% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{prRegisterTemplate}
\alias{prRegisterTemplate}
\title{Register a new script template}
\usage{
prRegisterTemplate(name, path, pattern = "", default = FALSE)
}
\arguments{
\item{name}{Name of the template. You can overhide a template by registering
a template with the same name.}

\item{path}{Path of the template file}

\item{pattern}{Regular expression. If user does not explicitely choose a 
template, but the script path matches this regular expression, then the
corresponding template is used.}

\item{default}{If `TRUE` then this template will be the default template}
}
\value{
This function is only used for side effects and does not return
  anything.
}
\description{
Script templates are used by function [prScript()] to create new scripts. 
This function can be used to register new script templates. It is 
especially useful to create custom project templates.

Template files are processed with function [brew::brew()], so you can 
use the `brew` syntax to insert some dynamic content in your scripts.
}
