% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{prStart}
\alias{prStart}
\title{Source start and tools scripts}
\usage{
prStart(dir = ".", trace = TRUE)
}
\arguments{
\item{dir}{Path of the directory where the project is stored.}

\item{trace}{Should the function print what it is doing ?}
}
\value{
TRUE if all scripts had been sourced without error and false otherwise.
}
\description{
This function sources all scripts prefixed by \code{tools} or situated in a 
subdirectory called \code{tools}, then it sources the \code{start.R} script.
It is automatically called when a project is opened.
}
\examples{
\dontrun{
prStart()
}

}
