% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requireVariable.R
\name{requireVariable}
\alias{requireVariable}
\title{Ask user the value of a variable}
\usage{
requireVariable(
  name,
  desc = NULL,
  default = NULL,
  what = character(),
  nmax = 1,
  alwaysAsk = TRUE,
  env = .GlobalEnv
)
}
\arguments{
\item{name}{Name of the required variable}

\item{desc}{Description to display}

\item{default}{Default value for the variable}

\item{what}{Type of data to read in interactive mode.}

\item{nmax}{Number of values to read in interactive mode.}

\item{alwaysAsk}{If the variable already exists, should the function ask a
new value?}

\item{env}{Environment where the variable should be defined}
}
\value{
Used for side effects
}
\description{
In interactive mode, this functions asks the user to type the value of a given
variable. In non-interactive mode, it searchs the value in the command line
arguments and it returns an error if the value does not exist.
}
