% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpMa.R
\name{addGitFolders}
\alias{addGitFolders}
\title{Extent folder list by git repository}
\usage{
addGitFolders(
  folders,
  git_repository = NULL,
  git_subfolders = NULL,
  lut_mode = FALSE
)
}
\arguments{
\item{folders}{list of subfolders within the project directory.}

\item{git_repository}{name of the project's git repository. Will be
added to the folders and subfolders defined in default "lut" or supplied by
user will be created.}

\item{git_subfolders}{subdirectories within git repository that should be
created.}

\item{lut_mode}{use predefined environmental settings. In this case, only the
name of the git repository must be supplied to the function.}
}
\description{
Extent folder list by git repository and create subdirectories
according to default values.
}
\examples{
\dontrun{

addGitFolders(folders = c("data", "data/tmp"), git_repository = "myproject")

}
}
\author{
Christoph Reudenbach, Thomas Nauss
}
\keyword{internal}
