% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{prScript}
\alias{prScript}
\title{create and/or open R scripts}
\usage{
prScript(name, template = c("analysis", "data", "function"), subdir = ".")
}
\arguments{
\item{name}{(optional) name of the script to create and/or open. This parameter can also include
the subfolder where to save the script (see examples).}

\item{template}{One of "analysis", "data" or "function". For new scripts, template to use.
prScript adds a few comments that encourage you to add comments and explain
what the script will do. These comments depend on the choosen template.}

\item{subdir}{subdirectory where the scripts needs to be created or opened. The 
subdirectory can also be directly specified in the parameter \code{name}.}
}
\description{
\code{prScript} creates and opens a new script for edition. If the script
already exists, it opens it in the Rstudio editor.
}
\details{
If the parameter \code{name} is missing, then a list of existing scripts is
displayed. The user is then invited to choose one by typing the number or 
name of the script he wants to open.
}
\examples{
\dontrun{

prScript("test")

prScript("myFunction", template = "function")

# Create script in a subfolder
prScript("test", subdir = "testdir")

# Or equivalently
prScript("testdir/test")
}

}
\seealso{
\code{\link{prSource}}, \code{\link{prMoveScript}}, \code{\link{prRenameScript}}, 
\code{\link{prDeleteScript}}
}

